Add stagecraft workers to api virtual machine.
Add three workers, 'worker', 'celerycam' and 'beat'.

We only run beat and celerycam on a single machine as multiple instances
of either of the workers would either cause duplicate periodic jobs to
get trigged or for the recording of the state of the job queue to get
done twice.

These workers have been broken out into individual classes as we will in
the future want to break workers out to their own machines and the beat
and cam classes needed to be linked to just a single machine.