Enable Skyscape off-site backup machine
As it stands, we back data up from the Skyscape production environment to
Carrenza since it is off-site from Skyscape. Now that we're moving
production to Carrenza, we need to update Hiera so that the new Carrenza
environment talks to the off-site Skyscape box. This does that by bringing
the backup Hiera keys in to production_migration since the Hiera hierarchy
is arranged such that this is looked at first before common.yaml where this
has typically been configured.