Add Licensify app in Signon to not update domain
We need a way to skip certain apps i.e. licensify 
as that is staying at the old domain name. The 
argument to skip apps is optional. If want
to skip an application the arguments are passed in
as so:

rake applications:migrate_domain['licensify, foo']

See: https://github.com/alphagov/signonotron2/pull/385