Force ntp Icinga checks to use IPv4.
There is now an IPv6 DNS entry for ntp.ubuntu.com.  This has caused lots
of the checks to fail with an error "can't create socket connection".
Investigating reveals that it's all the precise machines that have this.
On the trusty machines the check takes a long time, but eventually
works.

Adding the `-4` option to the check to force IPv4 fixes both problems,
and the check works quickly on both distros.