Add a grafana dashboard of 4XX codes
We have recently experienced two incidents

 - 2015-07-13: [Collections & Mainstream browse](https://gov-uk.atlassian.net/wiki/pages/viewpage.action?pageId=32604268)
 - 2015-06-26: [Whitehall 404 errors](https://gov-uk.atlassian.net/wiki/display/GOVUK/2015-06-26%3A+Whitehall+404+errors)

In these incidents, sections of GOV.UK started returning HTTP 404
responses for content which was previously valid. No automated
monitoring picked up these incidents.

We'd like to have automated monitoring if this happens in future. To
start making this process easier, we're adding this graph to improve the
visibility of these 4XX codes across the GOV.UK applications.

There's an RFC related to this PR here:
https://gov-uk.atlassian.net/wiki/display/FS/RFC+21:+Monitoring+for+404s

It's also worth saying that this graph uses a template over a static
file. This is despite not having an template variables in use. I chose
to do this to be consistent with the location of other graph configs.

Trello: https://trello.com/c/zd3HUNNn