Upgrade puppetlabs/stdlib to 4.8.0
This commit ugprades puppetlabs/stdlib to 4.8.0, which Puppet modules from
external sources installed by librarian-puppet are beginning to request when
the metadata.json (or equivalent) is parsed. An example is upgrading
alphagov/puppet-rcs to trail the latest commit to its master branch.

The changelog -
https://github.com/puppetlabs/puppetlabs-stdlib/blob/master/CHANGELOG.md -
has nothing alarming to me in it; we currently use 4.6.0, so changes between
that at 4.8.0 should be checked.