Add Icinga alert for errors at the CDN edge
We had a prolonged incident recently where our CDN provider was serving
with an elevated error rate because they had an issue with a cache node.

Our lack of monitoring meant that it took us a week to bring this to
our provider's attention, which in turn meant that we served far more
errors to users than we should have done.

This commit uses the metrics we are collecting about our CDN behaviour
and creates an Icinga alert if the error rate rises.

The warning and critical thresholds are deliberately low to begin with.
I don't believe that these values will alert excessively, but I have
absolutely no problem with raising them. The important thing is that
we have some visibility over how our CDN is behaving which we will
be able to act on quickly.

The check sets a `from` parameter to pull metrics from the last 30 minutes
because statistics from the CDN provider are delayed by several minutes.
This check is not intended to catch immediate issues - our internal alerting
will be far better at that. It's intended to catch subtle long-term issues
which we would not otherwise see.