Update icinga timeperiods
These haven't been updated in a long time and don't reflect the times
we actually expect in-office and on call support to work.

This commit updates the times to reflect when 2nd line is expected
to be in the office and renames the timeperiods. They're not currently
used anywhere but I intend to use them now.