Rewrite `icinga::pager_contact`
This commit rewrites our `icinga::pager_contact` defined type to
make it better reflect how we use it.

I've renamed `pager_contact` to `pagerduty_contact` because that's
what we actually use it for.

I've changed the interface to the defined type so that instead of
passing obscure Icinga options straight through, the defined type
does some work to make it obvious what the option is actually for.