Set up different alerts for in office and on call
We've had a lot of cases recently where PagerDuty phoned the on-call
support number because of a very short spike in errors. These errors
recovered quickly and there was no action to be taken.

This commit adds PagerDuty contacts that are specific to timeperiods.

This is so that the Nginx 5xx check on the cache machines can use different
thresholds for alerting depending on the time of day. I've done this
using very similar definitions of `icinga::check::graphite` because I think
that's the simplest way to get this working.

There's quite a bit of duplication in this commit which isn't ideal, but I'm
not sure there's a way to do this which will significantly reduce that.