Change line colour for Icinga Graphite links
This commit adds a `colorList` parameter to the Graphite render URL
which is a comma-separated list of colours to use on the graph.

I've moved the critical line and the warning line before the actual
target so that the order matches the colours. Red is critical, orange
is warning and the remaining colours are used for the metrics we're measuring.