Add Carrenza monitoring host entry for old staging.
Skycape staging needs to be able to resolve
monitoring/alert.cluster to Carrenza staging IP’s.

NRPE config has an `allowed_hosts` section, which
determines which hosts are able to connect to it.

Without this change NRPE only allows the old
monitoring machines (no longer in use) to connect
to it.