Add CPU model type Icinga check
Upon the migration to a new provider we discovered that some servers
weren't as performant upon hosts which had a different CPU model to what
they had previously. We cannot guarantee that if we bring up a fresh VM,
or if there is an issue and a VM is vmotioned to a new host that it'll
match the CPU we wish. This monitor will alert us if the defined CPU
model type is incorrect.

By default the CPU type check should be "intel" as that is what the
majority of our hosts have. A further commit will add the hieradata
classes for the VMs that require AMD.

Just to be clear, we do not have bias against AMD architecture, it's just the specific
hosts we were put on were not as performant compared to the Intel hosts
we have, and this is the easiest way to check.