Remove content check for Jenkins config.xml
Remove the two tests for the contents of the managed `config.xml` file.

We should have tests to verify that the contents are being set
correctly, but I don't think these tests are helpful as they don't check
which configuration parameter is being set; any occurrence of 'foo' or
'bar' in the file would cause the tests to pass.