Remove set -e from seed-crawler-wrapper
The seed-crawler-wrapper script handles errors by placing the exit
code from the seeder script into a variable called $CODE and then
checking this in the exit trap. The problem is that with set -e if the
sub-command fails then the script immediately exits before being able to
set $CODE.

Given that there are no subsequent commands that we really don't want to
run I think it's safe to remove the set -e here.