Don't use an asset CDN for whitehall admin
We use bootstrap to style our admin interfaces. By default bootstrap is
styled in a mobile up fashion. As older versions of IE ( IE <= 8 ) don't
understand media-queries by default they get mobile styling. So that our
admin interfaces don't look like mobile in these browsers we use
Respond.js which manually parses stylesheets and creates an appropriate
one for the current browser width.

IE 8 doesn't understand CORS headers, but the IE team did stop it from
making cross origin XHR requests. The way Respond.js works internally is
it makes an XHR request for the stylesheets. If they are on a different
domain (like a CDN) it can't get the response to parse. In the rest of
our admin interfaces this isn't a problem as we don't use a CDN for
assets. However, because whitehall has both the admin and the frontend
in the same application it is set up with the correct asset root. This
means all the admin assets are downloaded from the CDN.

This change sets the asset root env variable to an empty string on
whitehall admin machines, which will force it to get the assets locally
rather than from the CDN and behave like all our other admin
applications. This also then makes the admin interface appear with
multicolumn for our IE 8 users.