MongoDB Backup: Remove backup files older than 90 days
Remove latest, logs, daily, and weekly automongodbbackup files that are
older than 90 days.

We started to remove monthly files that were older than 90 days as part
of c649220 in #3320. This is an extension of that work.

There are currently backup files going back to 2013-11-30-06h37m in
production. We don't need to keep these around for so long.

Here is what one of the boxes looks like in production:

```
kushalp@production-mongo-2:~$ find /var/lib/automongodbbackup/{latest,logs,daily,weekly,monthly}/* -mtime +90
<SNIPPED OFF BIG LIST OF OUTPUT>
...
/var/lib/automongodbbackup/logs/ERRORS_127.0.0.1-999926097.log
/var/lib/automongodbbackup/weekly/week.39.2014-09-27_05h46m.tgz
/var/lib/automongodbbackup/weekly/week.48.2013-11-30_06h37m.tgz
/var/lib/automongodbbackup/weekly/week.48.2014-11-29_06h04m.tgz
```