All Carrenza Staging machines use Intel CPUs
Previously, machines in the new Carrenza Staging environment have been
running on AMD processors. This caused us performance problems, so we asked
Carrenza to migrate them wholesale to Intel processors which were generally
better behaved.

Icinga was set up to alert us if a machine was provisioned on a hypervisor
which only provided the wrong type of processor (for instance, if we were
expecting an Intel architecture, but the machine was provisioned on a
hypervisor that only provided AMD processors).

The use of Intel processors is the default in icinga::client::check_cputype,
so we can remove the overrides we were creating in Hiera here. Since there's
nothing else in these files, remove them.

This commit was automatically generated by a patch file, based on 5b27fd8.
Signed-off-by: Andrew Langhorn <andrew.langhorn@digital.cabinet-office.gov.uk>