Add Rake task to check syntax of Bash scripts
Add a Rake task to check the syntax of Bash scripts stored as templates
or files in our Puppet code base.

This task, which is included in the default Rake task, runs `bash -n`
against any templates or files that contain a hash-bang header (`#!`)
referring to bash and outputs any errors detected.

I have prioritised it before the `lint` task in the default Rake task as
this check is a lot faster, so Rake will report Bash syntax errors
sooner.

Note that this won't work for fragments of Bash scripts, or scripts
that don't use the hash-bang header.

When reading files, I use binary mode (`rb` flags passed to File#open)
to prevent a `invalid byte sequence in UTF-8` error when reading binary
files such as tarballs.

I did consider looking for templates or
files ending in `\.(ba)?sh(\.erb)?`, though there's a chance we'd match
a script using `sh` rather than `bash`, and empirical testing confirmed
that looking for the hash-bang header is sufficient.

Example output for a failing template and file:

    matt➜~/govuk/puppet(syntax_check_bash✗)» brake
    ---> Checking Puppetfile URLs
    ---> syntax:manifests
    ---> syntax:templates
    ---> syntax:hiera:yaml
    ---> Checking Bash script syntax
    rake aborted!
    modules/mongodb/templates/automongodbbackup:
    /var/folders/t_/0gc6ps9n5_q3_19q1bhk544r0000gn/T/puppet_bash_check20150907-28301-1f0x0h5:
    /var/folders/t_/0gc6ps9n5_q3_19q1bhk544r0000gn/T/puppet_bash_check20150907-28301-1f0x0h5: line 105: syntax error near unexpected token `done'
    /var/folders/t_/0gc6ps9n5_q3_19q1bhk544r0000gn/T/puppet_bash_check20150907-28301-1f0x0h5: line 105: `done'

    modules/govuk/files/foo:
    modules/govuk/files/foo: line 3: syntax error near unexpected token `done'
    modules/govuk/files/foo: line 3: `done'
    /Users/matt/govuk/puppet/lib/tasks/bash_syntax.rake:57:in `block in <top (required)>'
    Tasks: TOP => default => bash_syntax
    (See full trace by running task with --trace)