Allow MongoDB oplog size to be set
Allow the MongoDB oplog size to be set by passing a class parameter to
`mongo::server`. This will allow us to set the oplog size per replicaset
(cluster).

We've seen issues where MongoDB replicaset secondaries fail to keep up
with the primary's oplog when we sync data from our Production
environment to non-Production environments. The `errmsg` show by
`rs.status()` is:

    "still syncing, not yet to minValid optime 55ee580a:149f"

It appears that this is happening because the data sync drops all
existing databases and imports database dumps from Production, some of
which are significantly larger than the current oplog size.

The default oplog size is approximately 5% of available disk space [1]
on 64-bit Linux systems. On our boxen, this equates to approximately
3GB. Some of the databases being imported are as large as 35GB.

[1]:
http://docs.mongodb.org/v2.4/core/replica-set-oplog/#replica-set-oplog-sizing