Double oplog size on mongo and performance_mongo
Effectively double the oplog size on `mongo` and `performance_mongo`
machines by increasing it from approximately 3.5GB to 7GB.

I'm not 100% confident that this increase will be enough to remedy the
problems we're seeing, but we can increase it further later if
necessary.

We've seen issues where MongoDB replicaset secondaries fail to keep up
with the primary's oplog when we sync data from our Production
environment to non-Production environments. The `errmsg` show by
`rs.status()` is:

    "still syncing, not yet to minValid optime 55ee580a:149f"

It appears that this is happening because the data sync drops all
existing databases and imports database dumps from Production, some of
which are significantly larger than the current oplog size.

The default oplog size is approximately 5% of available disk space [1]
on 64-bit Linux systems. On our boxen, this equates to approximately
3.5GB. Some of the databases being imported are as large as 35GB.

[1]:
http://docs.mongodb.org/v2.4/core/replica-set-oplog/#replica-set-oplog-sizing