Change 2ndline dashboard to show large numbers
Replace the graphs in the 2ndline dashboard with large numbers
representing key metrics:

- HTTP requests per second at origin (as served from the cache nodes)
- HTTP 4XX error percentage at origin (as served from the cache nodes)
- HTTP 5XX error percentage at origin (as served from the cache nodes)
- HTTP 4XX error percentage served from edge for `www.gov.uk`,
  `redirector` (Bouncer) and `assets` Fastly services
- HTTP 5XX error percentage served from edge for `www.gov.uk`,
  `redirector` (Bouncer) and `assets` Fastly services
- HTTP requests per second at served from edge for `www.gov.uk`,
  `redirector` (Bouncer) and `assets` Fastly services

I think showing these numbers is much simpler to read, especially at a
distance. Members of the Second Line team rotate every week, so it's
very difficult to read a graph with a variable Y-axis from a distance
and identify any potential issues.

I've tried to keep the number of metrics shown here to a minimum.

Note that the metrics for origin don't include backend or Bouncer
traffic. Bouncer traffic is included in the edge metrics.

I've defined thresholds for each metric so that it will progressively
turn orange then red if the value hits a dangerous or critical
threshold. These will likely need tweaking over time.

Each metric can also be clicked to lead through to a more detailed
graph.