Specify movingAverage windowSize as quoted string
Graphite allows the `windowSize` argument to `movingAverage()` to be
specified as a quoted string rather than as a number of _N_ datapoints
[1].

Given that we record metrics at a 5-second interval, 12 datapoints are
equivalent to one minute, which is the value I've used in the quoted
string.

This change should make no functional difference but should make the
dashboards easier to read.

[1]: http://graphite.readthedocs.org/en/0.9.13-pre1/functions.html#graphite.render.functions.movingAverage