Set EXPOSE_GOVSPEAK for Smart Answers
To simplify coordination of content change requests between content designers
and developers we're going to expose Smart Answer outcomes as govspeak by
appending .txt to Smart Answer outcome URLs. This will allow content designers
to copy, paste and modify (in their text editor) govspeak. Currently they are
'reverse-engineering' it from HTML outputs.

We do not want to expose this feature to the wide public, but we want to enable
it on preview. The main concerns for not exposing it in production are:
- a fear of needing to support it in a long term
- it is an internal 'tool' and it feels clunky to have it available in
  production

Note that the values are strings 'true' or 'false', because it is an
environment variable and environment variables can only be strings.