Add fail2ban config to nginx module
Currently the fail2ban module adds nginx configuration even when a
machine does not have the nginx package installed. This causes fail2ban
to stall machines upon boot time as it is looking for nginx logfiles.

This change moves the fail2ban configuration over to the nginx module so
it is only added when nginx is configured.