Allow linking to Graphite from all Icinga checks
Links to Graphite were previously added in the `icinga::check::graphite`
defined type.

This commit adds a new parameter to the `icinga::check` defined type which
`icinga::check::graphite` uses. This parameter overrides the `action_url`
parameter to link to Graphite.

This can be used by Icinga alerts where it would be helpful to link
to metrics but don't actually alert based on the values of metrics.