Parameterise off-site backup monitoring constants
This commit parameterises the constants used in this manifest, since we
should be able to pass these in as class parameters now that there are
multiple off-site backup machines and different environments talk to
different off-site backup machines.

Set as class parameters so that we can set the values in Hiera.