Add monitoring for asset sync
The asset sync job had been failing for a couple of months and
we didn't notice because we don't monitor it at all.

This commit adds a passive monitoring check which will alert us
if the script fails to run successfully, or if the script doesn't
complete within the threshold (which is currently set to twice the
frequency that cron should run it).

I've only monitored one of the sync jobs because that should give
us enough confidence that the sync is working for all of them. The
script would need to change to accommodate multiple checks which I
think is more complex than it's worth.