Update carbon config file
This commit updates the `carbon.conf` file so that it mimics the
`carbon.conf.example` file which is installed. The Graphite docs
specify that the comments in this file are the canonical source
of documentation [1].

There are no functional changes to the config as a result of this commit.
The 3 properties which are changed are simply being set to their default
values.

[1]: http://graphite.readthedocs.org/en/latest/config-carbon.html#carbon-conf