Upgrade rspec to version 3.3
On master currently we have 617 spec examples. When run in parallel
in 8 processes having this number of examples causes the first block
of parallel tests (60 examples) to fail. Running with 4 processes
does not cause the tests to fail.

We could fix this by adding another test, but we'd run into this issue
again in the future.

Upgrading RSpec to 3.x also fixes the issue.

Using:
    bundle update rspec