Patch carbon-aggregator bug in Graphite 0.9.13
We recently upgraded to Graphite 0.9.13 which stopped
carbon-aggregator from generating aggregated metrics in Graphite.
We rely on these metrics for monitoring (eg 5xx status codes).

This bugfix is slated for release as part of 0.9.14, though we have
no guarantee about when that will be released.

This commit is very similar to (and borrows heavily from) Dan Carley's
work in 88e44970bdb804fe78856e61df268d20f8218a0c.