Remove hiera-puppet-helper
This commit removes hiera-puppet-helper in favour of the hiera
implementation in the newer version of rspec-puppet that we've
upgraded to.

I've had to change some tests to ensure they are reflective of
the common spec hieradata.

The govuk_mount specs require their own `hiera.yml` because they
explicitly test both cases of a `hiera()` lookup (of `govuk::mount::no_op`).