Import Transition site config only if repo has been updated
This job has been configured in deploy.production.alphagov.co.uk to run every
hour, but only if the release branch has been updated in the repo. When the
job was converted to use jenkins-job-builder that config was lost, and it has
been configured to run every hour regardless of whether the repo has changed.

This change should restore the previous behaviour. pollscm [1] is deprecated as of
jenkins-job-builder 1.3.0, but we're using 1.2.0.

'Build when a change is pushed to GitHub' has also been ticked as well as
'Poll SCM' under 'Build Triggers' in deploy.production.alphagov.co.uk, but
it's not clear to me from the documentation whether any additional config is
needed in the YAML file other than what I've added here to account for this.

[1]: http://docs.openstack.org/infra/jenkins-job-builder/triggers.html#triggers.pollscm