Enable loadbalancers to serve a maintenance page.
This change adds the ability to put a loadbalancer
into maintenance mode. It checks if maintenance_mode
is set to true and if so, writes a value of 1 to
`/etc/nginx/includes/maintenance.conf`. This
config file gets pulled in by nginx and serves
up the maintenance page if set to 1. It returns
a 503 when serving the maintenance page.
If statements are mostly evil, but OK in this
case.

The main reason for doing this now is so that during
migration, while signon isn't available we show a
maintenance page, and not nothing.