Convert spec to RSpec 3.3.2 syntax with Transpec
This spec test was on a branch when the rest of the tree was converted
to Rspec 3 syntax, and so was merged to master still using Rspec 2
syntax. This commit brings it in line with the rest of the tree per
3f0391a and d920e42.

This conversion is done by Transpec 3.1.1 with the following command:
    transpec modules/mongodb/spec/classes/mongodb__backup_spec.rb

* 5 conversions
    from: it { should ... }
      to: it { is_expected.to ... }

For more details: https://github.com/yujinakayama/transpec#supported-conversions

I then made some manual changes similar to d920e422d, where Transpec had
failed to correctly convert our syntax.

As part of this conversion, we no longer check for
`RSpec::Expectations::ExpectationNotMetError`, which we previously had
to use to assert that no error had been raised. Instead, it suffices to
assert that no error was raised.