Allow Carrenza Production to seed crawler worker
So that the content mirrored in Carrenza Production is up-to-date prior
to us enabling syncing of that content to the offsite mirrors, allow the
GOV.UK Crawler Worker to be seeded now.

So that we don't overwhelm the origin by hitting the long tail of
content from three different environments (Carrenza Production, Skyscape
Production and Carrenza Preview), disable seeding in Preview until the
migration is over.