Revert to master branch for Smokey
We configured Smokey to run from the `smokey-failure-licensing` branch
in Carrenza Production because Licensify was disabled in that
environment following an incident (see 9e091aa).

Now that we've migrated to Carrenza and we're going to re-enable
Licensify, switch back to the master branch for Smokey, which is the
default for this class parameter.