Change Errbit MongoDB replicaset members
617c506d configured our MongoDB replicasets such that backups only run
on one node in the set.

Puppet is currently erroring on `exception-hander-1.management` because
the hostname of the machine doesn't match any of the replicaset members
defined here in Hiera:

    Error: This machine's hostname was not found in the list of MongoDB replicaset
    members: localhost

This failure is part of a mechanism designed to ensure that if backups
are enabled, they always run on one node.

Fix the Puppet error by naming the replicaset member for Errbit after
the machine's hostname.