Fix ssh_known_hosts for new host naming.
The ssh module generates a global known_hosts file with the host keys of
all the machines in the same environment. This lists each hosts FQDN,
and IP, and in addition an 'internally qualified domain name' (IQDN).
This is intended to be of the form ${hostname}.${vdc}.

Currently, this is generated by removing the last component from the
FQDN. While this worked while we had FQDNs of the form
host-1.vdc.production, this no longer works with FQDNs of the form
host-1.vdc.publishing-service.gov.uk that we're using in the new
environments. As a result we end up with entries like the following in
the global known_hosts file:

api-mongo-1.api.publishing.service.gov.uk,api-mongo-1.api.publishing.service.gov,10.3.4.21
ssh-rsa AAAAB...

This makes it problematic when sshing between machines.

This PR therefore updates the regex to take the first to components of
the FQDN as the IQDN.

Note: it's not really possible to write a test of this because
rspec-puppet doesn't enable us to test exported resources. I've manually
tested the regex using rubular.com.  The results are here:
http://rubular.com/r/hGbc0oqjgu