Add handling for 12-factor style app logging.
In order to support the 12-factor-style logging scheme described in
RFC26[1], this updates `govuk::app` to conditionally collect these log
streams.

This introduces a new parameter `legacy_logging` to the type.  When set
to true, this makes govuk::app retain its current behaviour, otherwise
it assumes that the app logs JSON to stdout, and anything else to
stderr. This defaults to 'true' to ease the transition process. Once
the majority of apps have been updated to conform to the logging scheme
described in the RFC, this default can be revised.

[1]https://gov-uk.atlassian.net/wiki/display/TECH/RFC+26%3A+12-Factor+Rails+Apps