Fix pollscm `Transition_load_site_config` Jenkins
We're currently using jenkins-job-builder 1.2.0. The syntax for the
pollscm trigger changed in the most recent version (1.3.0) and the
docs on the site are only for the latest version.

Diff between 1.2.0 and current master: https://github.com/openstack-infra/jenkins-job-builder/compare/1.2.0...openstack-infra:b8eb602

Commit which changed pollscm: https://github.com/openstack-infra/jenkins-job-builder/commit/57d485e2a1633ca53313088acf69e899d28e826f

The syntax that I've moved us to in this commit is deprecated as of
the next version that we'll move to (1.3.0) but that's also the first
version where the new syntax appears.