Convert replicaset members to hash
We want to configure the priority of replicaset members in Puppet
which means converting the array of members to a hash.

This commit updates the places where we currently use the array of
replicaset members to only use the keys from the hash.

The Puppet to configure the priority will be added in a separate
commit.