Update Copy_Data_to_Staging job
This job was missing a couple of things:
 - Disabling and re-enabling Gor traffic replay
 - A separate shell execution for the Data_Sync_Complete job

The shell was also a little untidy so I have cleaned this up with some
line breaks etc (tested how JJB works with this and it's fine).

There was also an issue with `PLEASE CHOOSE ONE` incorrectly being
listed as an option where the default tried to run each of these words
as job. I have removed this line to ensure `all` is set as the default
job.