Add defined type for DATABASE_URL env var.
This is a wrapper around `govuk::app::envvar` that takes care of
constructing the url, and escaping the password.

When escaping the password, I looked at using the `uriescape` function
from stdlib, but that's not suitable in this instance. It's just a
wrapper around `URI.escape`, which doesn't cover many of the characters
we need in this instance. See
https://github.com/puppetlabs/puppetlabs-stdlib/pull/164 for more
details.

Instead, this uses an `inline_template` to call `CGI.escape`, which does
escape all the characters we need. It seemed overkill to create a custom
puppet function to do this.