Fix nodes spec after hiera-puppet-helper removal
Removing hiera-puppet-helper means that the value which is set as
`hiera_config` needs to be a path to a file rather than a YAML hierarchy.

Instead of duplicating our production `hiera.yml` as a spec fixture which
will have to be kept up-to-date constantly this commit creates a temporary
copy of that file with some values changed so that the tests are looking
for our hieradata in the correct places.