Switch `users::pentest_machines` to use hostname
The FQDN fact for these machines has changed in staging to be
`machine.vdc.staging.publishing.service.gov.uk` so this will
currently not working.

Switching to use the hostname facter fact rather than the fqdn facter
fact means that we don't need to include the environment of the
machines.