Licensify is external in Production
Licensify lives in Skyscape and the rest of the stack lives in Carrenza.
Plek was trying to infer the URI as living on a 'publishing.service.gov.uk'
URI, when in fact, it does live on a 'production.alphagov.co.uk' URI.
Override Plek using a PLEK_SERVICE_LICENSIFY_URI when $external_licensify is
set to 'true', which is set in Hiera (it defaults to false, since we want to
explicitly add this).

This commit is a partial revert of 45380d3.