Cleanup empty incoming asset dirs on asset-master.
The virus scanning script that runs from cron on the asset master, scans
all files in an incoming directory, and moves the clean ones to a clean
directory tree, and any infected ones to an infected directory tree.
This has the side effect of leaving lots of empty directories behind in
the incoming dir, which in turn has caused the scan process to gradually
take longer and longer because it has to scan through the many thousands
of empty directories looking for files.

This updates the scan script to remove these empty directories once the
scan has completed.