Add a Rake task to check YAML syntax
Add a Rake task to check the syntax of YAML stored in templates or files
in our Puppet code base.

This task, which is included in the default Rake task, will look for
files with a `.yaml` or `.yml` extension (also permitting a `.erb`
extension for ERB templates) and use the Ruby YAML library to
syntax-check YAML files and return any errors.

I have prioritised it before the lint task in the default Rake task as
this check is a lot faster, so Rake will report YAML syntax errors
sooner.

Note that this won't work for fragments of YAML files and could raise
false positives or negatives for certain ERB templates, though I think
this is an edge case.

Example output:
    $ brake yaml_syntax
    ---> Checking YAML file syntax
    rake aborted!
    modules/govuk_jenkins/templates/jobs/production/copy_data_to_preview.yaml.erb:
    (/var/folders/t_/0gc6ps9n5_q3_19q1bhk544r0000gn/T/puppet_yaml_lint20150928-13512-1570pn8): did not find expected key while parsing a block mapping at line 11 column 5
    /Users/matt/govuk/puppet/lib/tasks/yaml_syntax.rake:50:in `block in <top (required)>'
    Tasks: TOP => yaml_syntax
    (See full trace by running task with --trace)