Make jenkins_job_builder always update jobs.
jenkins_job_builder maintains a cache of the jobs it has written to
Jenkins. It uses this cache to determine whether a job needs to be
updated in Jenkins. This means that if a job has been edited in the
Jenkins UI, jenkins_job_builder won't notice, and the UI changes will
only be overwritten when there is a change to the job config in puppet..

Telling jenkins_job_builder to ignore it's cache means that it will
always write the jobs config into jenkins, and therefore any local
changes in the UI will be reverted on the next run.

For more details on this, see
https://bugs.launchpad.net/openstack-ci/+bug/1193444