Improve mirror sync script robustness.
Wrapping the rsync call in an if means that we can explicitly handle
failure without the script aborting. This means that if syncing to one
mirror fails, the script will still attempt to sync to the remaining
mirrors. If any of the rsync calls fail, this will still report an
overall failure to nagios.