Add Puppet to configure MongoDB node priority
We want to add a priority 0 MongoDB node in a different datacentre
for disaster recovery.

This commit adds a JavaScript template which will run on a MongoDB
primary to reconfigure a replicaset.

The reconfiguration will only happen if the values we specify in
Puppet are different from the values which are currently running
in the cluster.

Nodes have a priority of 1 by default and this value isn't returned
by the output of `rs.conf()`. The priority key is returned only when
the value isn't 1.