Create Jenkins Network Deploy jobs for Skyscape
In both Production and Staging environments we have machines running in
Skyscape for Licensify & EFG. This means we need to be able to push
firewall changes to these environments, but do not have a Jenkins
machine on these sides to track these changes. This change adds the jobs
to enable us to push these changes across.