Add govuk-cdn-logs-monitor app
Initially, this is only enabled in preview and vagrant environments,
while we test it.  It will eventually be enabled in all environments
other than development.

This will be deployed to the logs-cdn node, which means that node needs
the ruby app server adding to it.

I'm setting this up to place processed data from the app into the app's
`shared` directory.  I expect this to be of the order of 10Mb - if it
gets significantly larger, we can consider moving it to somewhere more
appropriate.

I've tested that this works appropriately with vagrant-govuk, but this
is the first app I've set up so may have done various things wrong.

See also https://github.gds/gds/alphagov-deployment/pull/1072