Add disaster recovery launch VMs job
This commit adds a Jenkins job to launch VMs in the disaster recovery
organisation.

Ideally we'd like to have this as a parameter in one Jenkins job:

- normal
- disaster recovery

But that actually ends up being quite complex to implement because the job
has to look up the value that's been selected and translate it to the
vCloud organisation and environment parameters.