Regroup Rake tasks to be logical
Previously, the `puppetfile` and `syntax` tasks were part of the
`default` task but not part of the `tests` task, which seems odd because
they are also tests.

The only useful distinction, as far as I can determine, is to separate
between the `lint` task and all other tests, because `jenkins.sh` needs
to run `lint` separately so as not to fail the build and prevent the
Jenkins lint plugin from parsing the results of the linter.

This simplifies the Jenkins CI configuration by ensuring that all tests
are run as part of the CI build. Previously, the `bash_syntax` and
`yaml_syntax` tasks were omitted from CI builds because I had wrongly
assumed that `jenkins.sh` would run all of the tests.