Change recurse parameter for Jenkins config
When this is set to `true` Puppet has to recursively gather the entire contents
of the `/var/lib/jenkins` directory which takes a very long time because it includes
all of our deployment jobs workspaces (which includes all of our source code) and
our Ruby bundles.

Setting the parameter to `'remote'` disables the `purge` option (which we're not
using) but speeds up the Puppet run to the point where it takes the same amount
of time that it did before adding the `recurse` param.

https://docs.puppetlabs.com/references/latest/type.html#file-attribute-recurse