Change Licensify hosts entries
Currently this is working for some reason which we're not sure about
(probably luck).

This commit removes the hosts entries for the Licensify applications
so that traffic for those apps goes over the internet rather than
over the VPN (making it more reliable).

However on the Licensify machines we need to set host entries for
the apps so that the requests are routed using the internal network.