Fix internal_tld on preview.
This was picking up the value from common.yaml and was therefore using
the production value (publishing.service.gov.uk). This has broken many
things including puppetdb which is expecting to listen on
puppetmaster-1.management.production, which no longer resolves.

This removes the value from common.yaml completely, as it's something
that's expected to vary per environment.