Pass paths to sync script as env variables
Following https://github.gds/gds/env-sync-and-backup/pull/152 (which has
been merged), we can pass database passwords to the env-sync-and-backup
scripts as environment variables.

This commit does this, for the affected databases (mysql and postgres),
instead of modifying the scripts to replace a PLACEHOLDER.

This will avoid writing our unencrypted root passwords to files.