Use new MapIt dump with Northern Ireland GSS code
While migrating business support schemes to use GSS codes instead of slugified
names to identify areas, we discovered that our MapIt database doesn't contain
a GSS code for the Northern Ireland European Region (EUR area type in MapIt).
This would mean that after the migration no results would be returned for
Northern Ireland postcodes in the business support finder.

This commit switches our existing MapIt machines to use a new dump of the
database which is the previous one with that GSS code added. I added the code
using a script [1]. This will allow us to continue with the migration.

I found that a few other areas in the currently active generation (2) are
missing codes in our database, but decided not to add them manually:

- all 26 Northern Ireland local authority areas have no GSS code: these areas
  ceased to exist on 1st April 2015 and were replaced by 11 local authorities.
  We do not have the boundary data for the new authorities and are unlikely to
  be able to find a source for the data, but as the ones we have in the
  database are out of date already there is no benefit in being able to refer
  to them via GSS codes. All NI business support schemes are already tagged to
  the EUR area and not the local authorities, so we don't need these codes for
  the migration.

- the Scotland and Northern Ireland EUR areas do not have SNAC codes ('ons' in
  MapIt): SNAC codes for these areas do exist, but are the same as two county
  councils in England (11 is Scotland EUR and Buckinghamshire CTY; 12 is
  Northern Ireland EUR and Cambridgeshire CTY). Adding these duplicate SNACs
  to our database could potentially break interactions on GOV.UK which rely on
  their uniqueness, and we don't know of anything which is broken now as a
  result of them being missing.

[1]: https://github.gds/gist/jennyduckett/2f41fdfd006ae9d72989